+++ CHANGELOG.md
@@ -5,11 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.2.0] - 2022-11-11
+## [0.1.5] - 2024-11-30
 ### Added
-- Method Sample.show_imp_ann_fracs() which outputs figures of implantation and annihilation fraction per layer vs. energy.
-- Sample.detailed_fit_result() now saves the figure to a file.
-- Output Sample.fit() result to text file.
+- Documentation for every class and function in limpid.py
+- Installation instructions for development.
+
+### Changed
+- Fix the handling of unsuccessful fits.
+- Move data/ folder into examples/ directory.
+- Version is now only defined in __init__.py. Setuptools can read it from there.
+- Fix a bug in the calculation of diffusion rates/fractions that would result in negative concentrations.
+
+### Removed
+- Deploy of branches other than main
 
 ## [0.1.1] - 2022-10-10
 ### Added